requirements.txt
PySocks==1.5.6
RPi.GPIO==0.6.1
httplib2==0.9.2
pytz==2015.7
six==1.10.0
twilio==5.3.0